feat: Remove AuthModel to only use OAuthModel
